# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FB Group F&B Platform API",
    "description": "Singapore F&B CRM / QR ordering / loyalty platform (PoC).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "List Campaigns",
        "operationId": "list_campaigns_api_v1_campaigns_get",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignListItemOut"
                  },
                  "title": "Response List Campaigns Api V1 Campaigns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign_api_v1_campaigns_post",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Campaign Detail",
        "operationId": "campaign_detail_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/audience": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Build Audience",
        "operationId": "build_audience_api_v1_campaigns__campaign_id__audience_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudienceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/send": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Send Campaign",
        "operationId": "send_campaign_api_v1_campaigns__campaign_id__send_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/redemptions": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Record Redemption",
        "operationId": "record_redemption_api_v1_campaigns__campaign_id__redemptions_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merchant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudienceResult": {
        "properties": {
          "audience_size": {
            "type": "integer",
            "title": "Audience Size"
          }
        },
        "type": "object",
        "required": [
          "audience_size"
        ],
        "title": "AudienceResult"
      },
...
      "CampaignCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "enum": [
              "whatsapp_promo",
              "birthday",
              "winback",
              "weekday_boost",
              "new_customer_return",
              "vip_reward"
            ],
            "title": "Campaign Type"
          },
          "segment_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Key"
          },
          "message_template": {
            "type": "string",
            "maxLength": 1000,
            "title": "Message Template",
            "default": ""
          },
          "reward_points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reward Points",
            "default": 0
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "campaign_type"
        ],
        "title": "CampaignCreateIn"
      },
      "CampaignDetailOut": {
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/CampaignOut"
          },
          "metrics": {
            "$ref": "#/components/schemas/CampaignMetricsOut"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Messages",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "campaign",
          "metrics"
        ],
        "title": "CampaignDetailOut"
      },
      "CampaignListItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "title": "Campaign Type"
          },
          "segment_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Key"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "metrics": {
            "$ref": "#/components/schemas/CampaignMetricsOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "campaign_type",
          "is_active",
          "created_at",
          "metrics"
        ],
        "title": "CampaignListItemOut"
      },
      "CampaignMetricsOut": {
        "properties": {
          "audience": {
            "type": "integer",
            "title": "Audience"
          },
          "sent": {
            "type": "integer",
            "title": "Sent"
          },
          "delivered": {
            "type": "integer",
            "title": "Delivered"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "redeemed": {
            "type": "integer",
            "title": "Redeemed"
          },
          "revenue_generated": {
            "type": "number",
            "title": "Revenue Generated"
          },
          "conversion_rate": {
            "type": "number",
            "title": "Conversion Rate"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "roi": {
            "type": "number",
            "title": "Roi"
          }
        },
        "type": "object",
        "required": [
          "audience",
          "sent",
          "delivered",
          "failed",
          "redeemed",
          "revenue_generated",
          "conversion_rate",
          "cost",
          "roi"
        ],
        "title": "CampaignMetricsOut"
      },
      "CampaignOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "title": "Campaign Type"
          },
          "segment_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Key"
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          },
          "reward_points": {
            "type": "integer",
            "title": "Reward Points"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "campaign_type",
          "message_template",
          "reward_points",
          "is_active",
          "created_at"
        ],
        "title": "CampaignOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "to_address": {
            "type": "string",
            "title": "To Address"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ref"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "to_address",
          "body",
          "status",
          "attempts",
          "created_at"
        ],
        "title": "MessageOut"
      },
...
      "RedemptionIn": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "revenue": {
            "type": "number",
            "minimum": 0.0,
            "title": "Revenue",
            "default": 0
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          }
        },
        "type": "object",
        "required": [
          "customer_id"
        ],
        "title": "RedemptionIn"
      },
...
      "SendResultOut": {
        "properties": {
          "delivered": {
            "type": "integer",
            "title": "Delivered"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "audience": {
            "type": "integer",
            "title": "Audience"
          }
        },
        "type": "object",
        "required": [
          "delivered",
          "failed",
          "audience"
        ],
        "title": "SendResultOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}